+++ Vortragsmanager/PageModels/MeineRednerKalenderPageModel.cs
@@ -19,6 +19,7 @@ namespace Vortragsmanager.PageModels
             ChangeView = new DelegateCommand<RednerViewType>(ChangeCurrentView);
             ListeSenden = new DelegateCommand(ListeVersenden);
             VortragAbsagen = new DelegateCommand(Absagen);
+            VortragBearbeiten = new DelegateCommand(Bearbeiten);
 
             var z = DataContainer.Redner.Where(x => x.Versammlung == DataContainer.MeineVersammlung);
             var enumerable = z.ToList();
@@ -57,6 +58,8 @@ namespace Vortragsmanager.PageModels
         public DelegateCommand<int> ChangeYear { get; }
 
         public DelegateCommand VortragAbsagen { get; }
+
+        public DelegateCommand VortragBearbeiten { get; }
 
         public void ChangeCurrentYear(int step)
         {
@@ -205,6 +208,35 @@ namespace Vortragsmanager.PageModels
                 DataContainer.ExternerPlan.Remove(SelectedTalk);
                 Talks.Remove(SelectedTalk);
             }
+        }
+
+        public void Bearbeiten()
+        {
+            var dlg1 = new KalendereintragVerschieben();
+            var data = (KalendereintragVerschiebenView)dlg1.DataContext;
+            var origKw = SelectedTalk.Kw;
+            data.LadeStartDatum(SelectedTalk);
+            dlg1.ShowDialog();
+
+
+            if (!data.Speichern) 
+                return;
+            
+            var origDataBaseItem = DataContainer.ExternerPlan.FirstOrDefault(x =>
+                x.Kw == origKw && x.Ältester.Name == data.StartName &&
+                x.Versammlung.Name == data.StartVersammlung);
+
+            if (origDataBaseItem != null)
+            {
+                origDataBaseItem.Kw = SelectedTalk.Kw;
+            }
+
+            if (data.ZielbuchungLöschenChecked && data.ZielBuchung != null)
+            {
+                Talks.Remove(data.ZielBuchung);
+            }
+
+            ApplyFilter();
         }
     }
 }