+++ Vortragsmanager/PageModels/WeekViewModel.cs
@@ -169,7 +169,6 @@ namespace Vortragsmanager.PageModels
         {
             var verschieben = new KalendereintragVerschieben();
             var data = (KalendereintragVerschiebenView)verschieben.DataContext;
-            data.KalenderTyp = Kalenderart.Intern;
             data.LadeStartDatum(Zuteilung);
             verschieben.ShowDialog();
 
@@ -194,6 +193,8 @@ namespace Vortragsmanager.PageModels
             var dialog = new RednerEintragenDialog();
             var data = (RednerEintragenView)dialog.DataContext;
             data.SelectedVersammlung = Einladung.Ältester.Versammlung;
+            dialog.InitializeSelectedConregation(Einladung.Ältester.Versammlung);
+            
             data.SelectedRedner = Einladung.Ältester;
             data.SelectedVortrag = Einladung.Vortrag;
             dialog.ShowDialog();