+++ Changelog.md
@@ -1,3 +1,7 @@
+### Version 1.1.22 (23.04.2022) ###
+* Externe Rednerbuchungen verschieben
+* kleinere Programmverbesserungen
+
 ### Version 1.1.18 (02.03.2022) ###
 * Neue Oberfläche, neue Navigation! Ich empfehle dir für die neue Navigation einen kurzen Blick ins Handbuch, Kapitel Startseite
 * Programm-Style umstellbar von Dunkel auf Hell